requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pdfplumber
pillow
pytesseract
lxml
pandas
python-dotenv
spacy
scikit-learn
regex
doccano-client   # opcional para anotações

# NOVAS DEPENDÊNCIAS PARA A SOLUÇÃO MELHORADA
opencv-python>=4.8.0        # Pré-processamento de imagens para OCR
python-dateutil>=2.8.0      # Manipulação de datas mais robusta
unidecode>=1.3.0           # Normalização de texto
numpy>=1.24.0              # Requisito do OpenCV e processamento